pyproject.toml
[tool.poetry]
name = "baboons-on-the-move"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
tqdm = "^4.66.1"
opencv-python = "^4.8.1.78"
opencv-contrib-python = "^4.8.1.78"
numpy = "^1.26.2"
pyaescrypt = "^6.1.1"
colorama = "^0.4.6"
sphinx = "^7.2.6"
pandas = "^2.1.4"
scikit-image = "^0.22.0"
matplotlib = "^3.8.2"
google-api-python-client = "^2.111.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = ">=0.5,<1.1"
firebase-admin = "^6.3.0"
numba = "^0.58.1"
pillow = "^10.1.0"
pyyaml = "^6.0.1"
synology-api = "^0.6.3"
jupyter = "^1.0.0"
gitpython = "^3.1.40"
py7zr = "^0.20.8"
protobuf = "^4.25.1"
sherlock = { git = "https://github.com/KastnerRG/sherlock.git", branch = "feature/poetry" }
bom-pipeline = {git = "https://github.com/UCSD-E4E/bom-pipeline", rev = "main"}
bom-spot = {git = "https://github.com/UCSD-E4E/bom-spot", rev = "main"}

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
pylint = "^3.0.3"
black = "^23.12.0"
ipykernel = "^6.27.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
